improve json-rpc error handling
